% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{csv_to_geojson_rust}
\alias{csv_to_geojson_rust}
\title{Convert \code{points.csv} to points_rust.geojson.}
\usage{
csv_to_geojson_rust()
}
\description{
Convert \code{points.csv} to points_rust.geojson.
}
