% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/csv_to_geojson.R
\name{csv_to_json_base_r}
\alias{csv_to_json_base_r}
\title{Convert csv file to geojson file}
\usage{
csv_to_json_base_r(file_csv, file_geojson = NULL)
}
\arguments{
\item{file_csv}{}

\item{file_geojson}{}
}
\value{

}
\description{
Convert csv file to geojson file
}
